relay/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
authlib>=1.3.0
httpx>=0.27.0
bcrypt>=4.1.0
aiosqlite>=0.20.0
itsdangerous>=2.1.0
python-dotenv>=1.0.0
starlette>=0.36.0
python-multipart>=0.0.9
